revise: 加上 404 錯誤提示; 驗證id格式; 正確、錯誤回應拉出來

+++ routes/posts.js
@@ -1,31 +1,23 @@
 var express = require('express');
 var router = express.Router();
 const Post = require('../models/postsModel')
+const checkIdFormat = require('../utils/checkFormat')
+const resHandle = require('../utils/resHandle')
 
 /* GET 取得全部貼文*/
 router.get('/', async function (req, res, next) {
   const posts = await Post.find()
-  res.status(200).json({
-    "status": "success",
-    "data": posts
-  })
+  resHandle.successHandle(res, posts)
 });
 
 /* POST 新增貼文*/
 router.post('/', async function (req, res, next) {
   try {
     const newPost = await Post.create(req.body)
-    res.status(200).json({
-      status: "success",
-      post: newPost
-    })
+    resHandle.successHandle(res, newPost)
   } catch (error) {
     let errorMessage = Object.values(error.errors).map(item => item.message)
-    res.status(400).json({
-      status: "false",
-      message: "欄位格式錯誤",
-      error: errorMessage
-    })
+    resHandle.errorHandle(res, 400, errorMessage)
   }
 });
 
@@ -37,28 +29,23 @@ router.delete('/:id?', async function (req, res, next) {
   */
   if (!req.params.id) {
     await Post.deleteMany({})
-    res.status(200).json({
-      status: "success",
-      post: []
-    })
+    resHandle.successHandle(res)
   } else {
+    // 驗證 ID 格式是否正確
+    if (!checkIdFormat.checkIdFormat(req.params.id)) {
+      resHandle.errorHandle(res, 400, ['查無此 ID'])
+      return
+    }
+
     // 查詢是否有此 ID 的資料
     const hasId = await Post.findById(req.params.id)
 
     if (hasId) {
       await Post.findByIdAndDelete(req.params.id)
       const posts = await Post.find()
-
-      res.status(200).json({
-        status: "success",
-        post: posts
-      })
+      resHandle.successHandle(res, posts)
     } else {
-      res.status(400).json({
-        status: "false",
-        message: "欄位格式錯誤",
-        error: ['查無此 ID']
-      })
+      resHandle.errorHandle(res, 400, ['查無此 ID'])
     }
   }
 });
@@ -66,33 +53,29 @@ router.delete('/:id?', async function (req, res, next) {
 /* PATCH 編輯貼文*/
 router.patch('/:id', async function (req, res, next) {
   try {
-    // 先確認是否有資料
+    // 驗證 ID 格式是否正確
+    if (!checkIdFormat.checkIdFormat(req.params.id)) {
+      resHandle.errorHandle(res, 400, ['查無此 ID'])
+      return
+    }
+
+    // 確認是否有資料
     const hasId = await Post.findById(req.params.id)
     if (req.params.id && hasId) {
 
       // 若無填寫 content，則回傳錯誤
       if (!req.body.content) {
-        res.status(400).json({
-          status: "false",
-          message: "欄位格式錯誤",
-          error: ['貼文內容 content 未填寫']
-        })
+        resHandle.errorHandle(res, 400, ['貼文內容 content 未填寫'])
         return
       }
       await Post.findByIdAndUpdate(req.params.id, req.body)
       const post = await Post.findById(req.params.id)
-      res.status(200).json({
-        "status": "success",
-        "data": post
-      })
+      resHandle.successHandle(res, post)
     }
   } catch (error) {
-    res.status(400).json({
-      status: "false",
-      message: "欄位格式錯誤",
-      error: ['查無此 ID']
-    })
+    resHandle.errorHandle(res, 400, error)
   }
 });
+
 
 module.exports = router;
